Make sure tests use translated repeat language
