+++ collect_app/src/androidTest/java/org/odk/collect/android/regression/FillBlankFormTest.java
@@ -151,7 +151,7 @@ public class FillBlankFormTest extends BaseRegressionTest {
 
         //TestCase41
         new MainMenuPage(rule)
-                .startBlankFormWithRepeatGroup("formulaire_adherent")
+                .startBlankFormWithRepeatGroup("formulaire_adherent", "Ajouté une observation")
                 .clickOnAddGroup(new FormEntryPage("formulaire_adherent", rule))
                 .clickOnText("Plante")
                 .inputText("Abi")
