--- collect_app/src/androidTest/java/org/odk/collect/android/support/pages/AddNewGroupDialog.java
@@ -1,29 +1,0 @@
-package org.odk.collect.android.support.pages;
-
-import androidx.test.rule.ActivityTestRule;
-
-import org.odk.collect.android.R;
-
-public class AddNewGroupDialog extends Page<AddNewGroupDialog> {
-
-    AddNewGroupDialog(ActivityTestRule rule) {
-        super(rule);
-    }
-
-    @Override
-    public AddNewGroupDialog assertOnPage() {
-        checkIsStringDisplayed(R.string.entering_repeat_ask);
-        return this;
-    }
-
-    public FormEntryPage clickOnAddGroup(FormEntryPage destination) {
-        clickOnString(R.string.add_repeat);
-        return destination;
-    }
-
-    public FormEntryPage clickOnDoNotAddGroup(FormEntryPage destination) {
-        clickOnString(R.string.add_repeat_no);
-        return destination;
-    }
-
-}
