+++ collect_app/src/main/res/values-af/strings.xml
@@ -41,8 +41,7 @@
   <string name="do_not_change">Kanselleer</string>
   <string name="do_not_exit">Kanselleer</string>
   <string name="do_not_save">Ignoreer Veranderinge</string>
-    <string name="entering_repeat_ask">Voeg Nuwe Groep By?</string>
-  <string name="enter_data">Vul Leë Vorm In</string>
+    <string name="enter_data">Vul Leë Vorm In</string>
   <string name="enter_data_button">Vul Leë Vorm In</string>
   <string name="error_occured">Daar was \'n fout</string>
   <string name="error_attaching_binary_file">\'n Fout het gebeur tydens aanheg van media: %s</string>
