+++ collect_app/src/main/res/values-am/strings.xml
@@ -40,7 +40,6 @@
   <string name="do_not_change">ሰርዝ</string>
   <string name="do_not_exit">ሰርዝ</string>
   <string name="do_not_save">ለውጦቹን ችላ በል</string>
-    <string name="entering_repeat_ask">አዲስ ቡድን ይጨመር?</string>
   <string name="enter_data">ባዶዉን ቅጽ ሙላ</string>
   <string name="enter_data_button">ባዶ ቅጽ መሙላት</string>
   <string name="error_occured">ስህተት ተከስቷል</string>
