+++ collect_app/src/main/res/values-in/strings.xml
@@ -51,8 +51,7 @@
   <string name="do_not_change">Batal</string>
   <string name="do_not_exit">Batal</string>
   <string name="do_not_save">Abaikan perubahan</string>
-    <string name="entering_repeat_ask">Tambah grup baru?</string>
-  <string name="enter_data">Isi formulir kosong</string>
+    <string name="enter_data">Isi formulir kosong</string>
   <string name="enter_data_button">Isi formulir kosong</string>
   <string name="error_occured">Terjadi kesalahan</string>
   <string name="error_attaching_binary_file">Terjadi kesalahan saat memasukkan berkas media: %s</string>
