+++ collect_app/src/main/res/values-ja/strings.xml
@@ -51,8 +51,7 @@
   <string name="do_not_change">キャンセル</string>
   <string name="do_not_exit">キャンセル</string>
   <string name="do_not_save">変更を無視する</string>
-    <string name="entering_repeat_ask">新しいグループを追加しますか?</string>
-  <string name="enter_data">空のフォームに入力してください</string>
+    <string name="enter_data">空のフォームに入力してください</string>
   <string name="enter_data_button">空のフォームに入力する</string>
   <string name="error_occured">エラーが発生しました</string>
   <string name="error_attaching_binary_file">メディアの添付中にエラーが発生しました: %s</string>
