+++ collect_app/src/main/res/values-ka/strings.xml
@@ -51,8 +51,7 @@
   <string name="do_not_change">გაუქმება</string>
   <string name="do_not_exit">გაუქმება</string>
   <string name="do_not_save">ცვლილებების გაუქმება</string>
-    <string name="entering_repeat_ask">დავამატოთ ახალი ჯგუფი?</string>
-  <string name="enter_data">ცარიელი ფორმის შევსება</string>
+    <string name="enter_data">ცარიელი ფორმის შევსება</string>
   <string name="enter_data_button">შეავსეთ ცარიელი ფორმა</string>
   <string name="error_occured">მოხდა შეცდომა</string>
   <string name="error_attaching_binary_file">შეცდომა მოხდა მედიაფაილის მიმაგრებისას: %s</string>
