+++ collect_app/src/main/res/values-my/strings.xml
@@ -50,8 +50,7 @@
   <string name="do_not_change">မလုပ်တော့ပါ</string>
   <string name="do_not_exit">မလုပ်တော့ပါ</string>
   <string name="do_not_save">ပြင်ဆင်မှုများကို လက်မခံပါ</string>
-    <string name="entering_repeat_ask">အုပ်စုအသစ် ထည့်သွင်းလိုပါသလား</string>
-  <string name="enter_data">ပုံစံအလွတ်အား ဖြည့်သွင်းရန်</string>
+    <string name="enter_data">ပုံစံအလွတ်အား ဖြည့်သွင်းရန်</string>
   <string name="enter_data_button">ပုံစံအလွတ်အား ဖြည့်သွင်းရန်</string>
   <string name="error_occured">မွားယြင္းေနပါသည္</string>
   <string name="file_deleted_error">%1$s ၊ %2$s ေရြးခ်ယ္ထားေသာပံုစံမ်ားအားဖ်က္၍ မရပါ</string>
