+++ collect_app/src/main/res/values-nl/strings.xml
@@ -51,8 +51,7 @@
   <string name="do_not_change">Annuleer</string>
   <string name="do_not_exit">Annuleer</string>
   <string name="do_not_save">Wijzigingen niet opslaan</string>
-    <string name="entering_repeat_ask">Nieuwe groep toevoegen?</string>
-  <string name="enter_data">Nieuw formulier invullen</string>
+    <string name="enter_data">Nieuw formulier invullen</string>
   <string name="enter_data_button">Nieuw formulier</string>
   <string name="error_occured">Fout opgetreden</string>
   <string name="error_attaching_binary_file">Er is een fout opgetreden bij het koppelen van media: %s</string>
