+++ collect_app/src/main/res/values-no/strings.xml
@@ -41,8 +41,7 @@
   <string name="do_not_change">Avbryt</string>
   <string name="do_not_exit">Avbryt</string>
   <string name="do_not_save">Annuler endringer</string>
-    <string name="entering_repeat_ask">Legge til ny gruppe?</string>
-  <string name="enter_data">Fyll ut skjema</string>
+    <string name="enter_data">Fyll ut skjema</string>
   <string name="enter_data_button">Fyll ut skjema</string>
   <string name="error_occured">Det er oppstatt en feil</string>
   <string name="error_attaching_binary_file">En feil oppstod under lagring av media: %s</string>
