+++ collect_app/src/main/res/values-pt/strings.xml
@@ -44,8 +44,7 @@
   <string name="do_not_change">Cancelar</string>
   <string name="do_not_exit">Cancelar</string>
   <string name="do_not_save">Ignorar Alterações</string>
-    <string name="entering_repeat_ask">Adicionar Novo Grupo?</string>
-  <string name="enter_data">Formulário em Branco</string>
+    <string name="enter_data">Formulário em Branco</string>
   <string name="enter_data_button">Formulário em Branco</string>
   <string name="error_occured">Ocorreu um erro</string>
   <string name="error_attaching_binary_file">Um erro ocorreu ao anexar o arquivo: %s</string>
