+++ collect_app/src/main/res/values-ro/strings.xml
@@ -41,8 +41,7 @@
   <string name="do_not_change">Anulați</string>
   <string name="do_not_exit">Anulați</string>
   <string name="do_not_save">Ignorați modificările</string>
-    <string name="entering_repeat_ask">Adăugați un grup nou?</string>
-  <string name="enter_data">Completați formular necompletat</string>
+    <string name="enter_data">Completați formular necompletat</string>
   <string name="enter_data_button">Completați formular necompletat</string>
   <string name="error_occured">Eroare</string>
   <string name="error_attaching_binary_file">A intervenit o eroare în timpul atașării fișierului media: %s</string>
