+++ collect_app/src/main/res/values-so/strings.xml
@@ -17,8 +17,7 @@
   <string name="discard_group">Ka saar kooxda</string>
   <string name="do_not_change">Jooji</string>
   <string name="do_not_exit">Jooji</string>
-    <string name="entering_repeat_ask">Kudar koox cusub</string>
-  <string name="enter_data">Buuxi foomka bannaan</string>
+    <string name="enter_data">Buuxi foomka bannaan</string>
   <string name="enter_data_button">Buuxi foomka bannaan</string>
   <string name="error_occured">Cilad ku timi</string>
   <string name="forms">Foomamka Banaan</string>
