+++ collect_app/src/main/res/values-th-rTH/strings.xml
@@ -41,8 +41,7 @@
   <string name="do_not_change">ยกเลิก</string>
   <string name="do_not_exit">ยกเลิก</string>
   <string name="do_not_save">ยกเลิกการเปลี่ยนแปลง</string>
-    <string name="entering_repeat_ask">เพิ่มกลุ่มใหม่</string>
-  <string name="enter_data">กรอกแบบสำรวจใหม่</string>
+    <string name="enter_data">กรอกแบบสำรวจใหม่</string>
   <string name="enter_data_button">กรอกแบบสำรวจใหม่</string>
   <string name="error_occured">มีข้อผิดพลาดเกิดขึ้น</string>
   <string name="error_attaching_binary_file">เกิดข้อผิดพลาดระหว่างแนบไฟล์มีเดีย: %s</string>
