+++ collect_app/src/main/res/values-tl-rPH/strings.xml
@@ -39,8 +39,7 @@
   <string name="do_not_change">Urong</string>
   <string name="do_not_exit">Urong</string>
   <string name="do_not_save">Balewalain ang mga binago</string>
-    <string name="entering_repeat_ask">Mag-dagdag ng Bagong Pangkat?</string>
-  <string name="enter_data">Punan ang Blangkong Form</string>
+    <string name="enter_data">Punan ang Blangkong Form</string>
   <string name="enter_data_button">Punan ang Form</string>
   <string name="error_occured">May Maling Nangyari</string>
   <string name="file_deleted_error">Paumanhin, %1$s of %2$s ang mga piniling form(s) ay nabigong ma-i-karga!</string>
