+++ collect_app/src/main/res/values-uk/strings.xml
@@ -51,8 +51,7 @@
   <string name="do_not_change">Скасувати</string>
   <string name="do_not_exit">Скасувати</string>
   <string name="do_not_save">Не зберігати</string>
-    <string name="entering_repeat_ask">Додати нову групу?</string>
-  <string name="enter_data">Заповнити анкету</string>
+    <string name="enter_data">Заповнити анкету</string>
   <string name="enter_data_button">Заповнити анкету</string>
   <string name="error_occured">Сталась помилка</string>
   <string name="error_attaching_binary_file">При додаванні медіафайлу: %s сталась помилка</string>
