+++ collect_app/src/main/res/values-ur-rPK/strings.xml
@@ -41,8 +41,7 @@
   <string name="do_not_change">منسوخ</string>
   <string name="do_not_exit">منسوخ</string>
   <string name="do_not_save">تبدیلیاں نظرانداز کریں</string>
-    <string name="entering_repeat_ask">نیا گروپ شامل کریں</string>
-  <string name="enter_data">خالی فارم پر کریں</string>
+    <string name="enter_data">خالی فارم پر کریں</string>
   <string name="enter_data_button">خالی فارم پر کریں</string>
   <string name="error_occured">غلطی ہوگئ</string>
   <string name="fetching_file">‎%1$s‏ کو حاصل کیا جا رہا ہے؟؟ فارم (ز) ‎%3$s‏ کا فارم ‎%2$s‏</string>
