+++ collect_app/src/main/res/values/strings.xml
@@ -61,7 +61,6 @@
     <string name="do_not_change">Cancel</string>
     <string name="do_not_exit">Cancel</string>
     <string name="do_not_save">Ignore Changes</string>
-    <string name="entering_repeat_ask">Add New Group?</string>
     <string name="enter_data">Fill Blank Form</string>
     <string name="enter_data_button">Fill Blank Form</string>
     <string name="error_occured">Error Occurred</string>
